[BACKLOG-14117]-Create and validate an Azure shim
