+++ hdi35/build.properties
@@ -1,0 +1,39 @@
+impl.title=Pentaho Hadoop HDI 3.5.x Shim
+impl.productID=pentaho-hadoop-shims-hdi35
+ivy.artifact.id=pentaho-hadoop-shims-hdi35
+ivy.artifact.group=pentaho
+project.revision=7.1-SNAPSHOT
+package.id=pentaho-hadoop-shims-hdi35-package
+package.root.dir=hdi35
+oss-license.filename=PentahoHadoopShim_hdi35_OSS_Licenses.html
+
+# Common parts integrating
+common.modern.shim.include=true
+common.hadoop.shim.version=1.0
+common.hbase.shim.version=1.1
+common.pig.shim.version=1.1
+hbase.generation=1.0
+
+dependency.kettle.revision=7.1-SNAPSHOT
+dependency.hadoop-shims-api.revision=7.1-SNAPSHOT
+dependency.pentaho-hdfs-vfs.revision=7.1-SNAPSHOT
+dependency.oss-licenses.revision=7.1-SNAPSHOT
+
+dependency.commons-lang.revision=2.6
+
+dependency.hadoop.revision=2.7.3.2.5.0.0-1245
+dependency.hadoop2-windows-patch.revision=08072014
+dependency.hive-jdbc.revision=1.2.1000.2.5.0.0-1245
+dependency.apache-hbase.revision=1.1.2.2.5.0.0-1245
+dependency.apache-sqoop.revision=1.4.6.2.5.0.0-1245
+dependency.pig.revision=0.16.0.2.5.0.0-1245
+dependency.automaton.revision=1.11-8
+dependency.zookeeper.revision=3.4.6.2.5.0.0-1245
+dependency.apache-oozie.revision=4.2.0.2.5.0.0-1245
+protobuf-java.version=2.5.0
+dependency.metrics-core.revision=2.2.0
+
+junit.sysprop.org.pentaho.hadoop.shims.check.hive1=false
+
+izpack.version=5.0.7-RC1
+pentaho-eula-wrap-config.version=1.0.10
