+++ hdi35/package-res/config.properties
@@ -1,0 +1,42 @@
+#
+#  PENTAHO CORPORATION PROPRIETARY AND CONFIDENTIAL
+#
+#  Copyright 2002 - 2016 Pentaho Corporation (Pentaho). All rights reserved.
+#
+#  NOTICE: All information including source code contained herein is, and
+#  remains the sole property of Pentaho and its licensors. The intellectual
+#  and technical concepts contained herein are proprietary and confidential
+#  to, and are trade secrets of Pentaho and may be covered by U.S. and foreign
+#  patents, or patents in process, and are protected by trade secret and
+#  copyright laws. The receipt or possession of this source code and/or related
+#  information does not convey or imply any rights to reproduce, disclose or
+#  distribute its contents, or to manufacture, use, or sell anything that it
+#  may describe, in whole or in part. Any reproduction, modification, distribution,
+#  or public display of this information without the express written authorization
+#  from Pentaho is strictly prohibited and in violation of applicable laws and
+#  international treaties. Access to the source code contained herein is strictly
+#  prohibited to anyone except those individuals and entities who have executed
+#  confidentiality and non-disclosure agreements or other agreements with Pentaho,
+#  explicitly covering such access.
+#
+# Friendly name for this configuration
+name=Microsoft Azure HDI 3.5.x
+
+# Comma-separated list of directories and files to make available to this
+# configuration. Any resources found here will overwrite ones in lib/.
+classpath=lib/hadoop2-windows-patch-08072014.jar
+
+# Comma-separated list of paths that contain native libraries to load. These
+# could be added to LD_LIBRARY_PATH or set with -Djava.library.path instead.
+library.path=
+
+# Comma-separated list of classes or package names to explicitly ignore when
+# loading classes from the resources within this Hadoop configuration directory
+# or the classpath property
+# e.g.: org.apache.commons.log,org.apache.log4j
+# Note, the two packages above are automatically included for all configurations
+ignore.classes=java.security.Permission,org.apache.derby
+
+java.system.hdp.version=2.5.2.1-1
+
+scheme=wasb://clusterName@storageName.blob.core.windows.net/
