+++ hdi35/package-res/hbase-site.xml
@@ -1,0 +1,7 @@
+<!--Thu May 30 10:40:32 2013-->
+  <configuration>
+    <property>
+      <name>zookeeper.znode.parent</name>
+      <value>/hbase-unsecure</value>
+    </property>
+  </configuration>
