+++ hdi35/package-res/mapred-site.xml
@@ -1,0 +1,57 @@
+<!--
+  - These settings are the minimum necessary MapReduce settings for the Hortonworks 2.0 Sandbox. The settings for your cluster may be different,
+  - please see the descriptions below along with your vendor documentation for the appropriate property values.
+  -
+  - The Apache defaults (not necessarily vendor-compatible) are listed here:
+  - http://hadoop.apache.org/docs/current/hadoop-mapreduce-client/hadoop-mapreduce-client-core/mapred-default.xml
+  -->
+<configuration>
+  <!--
+    - The following property specifies application classpath.
+    -->
+  <property>
+    <name>mapreduce.application.classpath</name>
+    <value>$PWD/mr-framework/hadoop/share/hadoop/mapreduce/*:$PWD/mr-framework/hadoop/share/hadoop/mapreduce/lib/*:$PWD/mr-framework/hadoop/share/hadoop/common/*:$PWD/mr-framework/hadoop/share/hadoop/common/lib/*:$PWD/mr-framework/hadoop/share/hadoop/yarn/*:$PWD/mr-framework/hadoop/share/hadoop/yarn/lib/*:$PWD/mr-framework/hadoop/share/hadoop/hdfs/*:$PWD/mr-framework/hadoop/share/hadoop/hdfs/lib/*:/usr/hdp/${hdp.version}/hadoop/lib/hadoop-lzo-0.6.0.${hdp.version}.jar:/etc/hadoop/conf/secure</value>
+  </property>
+
+  <!--
+    - The following property specifies mapreduce framework path.
+    -->
+  <property>
+    <name>mapreduce.application.framework.path</name>
+    <value>/hdp/apps/${hdp.version}/mapreduce/mapreduce.tar.gz#mr-framework</value>
+  </property>
+
+  <!--
+    - The following property specifies in which mode to run MapReduce jobs.
+    - the choices are local, classic, or yarn:
+    -
+    - local: run the MapReduce tasks on this client machine
+    - classic: run job(s) on the cluster using MapReduce v1
+    - yarn: run job(s) on the cluster using MapReduce v2 (on YARN)
+    -
+    -->
+  <property>
+    <name>mapreduce.framework.name</name>
+    <value>yarn</value>
+  </property>
+
+  <!--
+    - The following property allows to run MR jobs cross-platformly.
+    - It means that jobs created on Windows MR client can be run on Linux cluster or vice versa.
+    -
+    -->
+  <property>
+    <name>mapreduce.app-submission.cross-platform</name>
+    <value>true</value>
+  </property>
+
+  <!--
+    - This property specifies the hostname:port value for the MapReduce JobHistory Server of your cluster/sandbox. Update this property accordingly
+    -->
+  <property>
+    <name>mapreduce.jobhistory.address</name>
+    <value>sandbox.hortonworks.com:10020</value>
+  </property>
+
+</configuration>
