+++ hdi35/src/META-INF/services/org.pentaho.hadoop.shim.common.authorization.NoOpHadoopAuthorizationService
@@ -1,0 +1,1 @@
+org.pentaho.hadoop.shim.hdi35.authorization.ShimNoOpHadoopAuthorizationService
